backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
transformers>=4.40.0
torch==2.6.0
firebase-admin==6.3.0
PyPDF2==3.0.1
pdfminer.six==20231228
langdetect==1.0.9
numpy>=1.26.0
pandas>=2.1.3
scikit-learn>=1.3.2
gunicorn==21.2.0
google-generativeai>=0.3.2
